forwardRef 삭제 및 공통 컴포넌트 분리
