+++ src/shared/ui/Input.tsx
@@ -1,0 +1,13 @@
+import type { ElementBaseProps } from "./type"
+
+type Props<T> = React.InputHTMLAttributes<T> & ElementBaseProps<T>
+
+export const Input = ({ className, ...props }: Props<HTMLInputElement>) => {
+  return (
+    <input
+      className={`flex h-10 w-full rounded-md border border-input bg-white px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 ${className}`}
+      {...props}
+    />
+  )
+}
+Input.displayName = "Input"
