+++ src/shared/ui/type.ts
@@ -1,0 +1,1 @@
+export type ElementBaseProps<T> = React.DetailedHTMLProps<React.HTMLAttributes<T>, T>
